Separated notebook and cell utility modules into separate files.
